requirements.txt
# Core data science and machine learning packages
numpy>=1.25
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Visualization packages
matplotlib==3.7.2
seaborn==0.12.2

# Flask backend packages
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Image processing
Pillow==10.0.0

# Development tools
pytest==7.4.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
